# Conflicts:
#	app/Http/Controllers/Admin/Characters/CharacterController.php
#	app/Http/Controllers/Admin/Characters/CharacterImageController.php
#	app/Http/Controllers/BrowseController.php
#	app/Models/Character/CharacterDesignUpdate.php
#	app/Models/Character/CharacterImage.php
#	app/Services/CharacterManager.php
#	config/lorekeeper/extension_tracker.php
#	resources/views/admin/items/tags/slot.blade.php
#	resources/views/admin/masterlist/_create_character_subtype.blade.php
#	resources/views/admin/masterlist/create_character.blade.php
#	resources/views/character/_image_info.blade.php
#	resources/views/character/admin/_edit_features_modal.blade.php
#	resources/views/character/admin/_edit_features_subtype.blade.php
#	resources/views/character/admin/_upload_image_subtype.blade.php
#	resources/views/character/admin/upload_image.blade.php
#	resources/views/character/design/features.blade.php
#	routes/lorekeeper/admin.php

+++ app/Services/CharacterManager.php
@@ -109,6 +109,13 @@ class CharacterManager extends Service
                 if(!$subtype || $subtype->species_id != $data['species_id']) throw new \Exception('Selected subtype invalid or does not match species.');
             }
             else $data['subtype_id'] = null;
+            if(isset($data['subtype_id_2']) && $data['subtype_id_2'])
+            {
+                $subtypeTwo = Subtype::find($data['subtype_id_2']);
+                if(!(isset($data['species_id']) && $data['species_id'])) throw new \Exception('Species must be selected to select a subtype.');
+                if(!$subtypeTwo || $subtype->species_id != $data['species_id']) throw new \Exception('Selected secondary subtype invalid or does not match species.');
+            }
+            else $data['subtype_id_2'] = null;
 
             // Get owner info
             $url = null;
@@ -204,6 +211,7 @@ class CharacterManager extends Service
                 $data['slug'] = null;
                 $data['species_id'] = isset($data['species_id']) && $data['species_id'] ? $data['species_id'] : null;
                 $data['subtype_id'] = isset($data['subtype_id']) && $data['subtype_id'] ? $data['subtype_id'] : null;
+                $data['subtype_id_2'] = isset($data['subtype_id_2']) && $data['subtype_id_2'] ? $data['subtype_id_2'] : null;
                 $data['rarity_id'] = isset($data['rarity_id']) && $data['rarity_id'] ? $data['rarity_id'] : null;
                 $data['transformation_id'] = isset($data['transformation_id']) && $data['transformation_id'] ? $data['transformation_id'] : null;
             }
@@ -254,6 +262,7 @@ class CharacterManager extends Service
             {
                 $data['species_id'] = (isset($data['species_id']) && $data['species_id']) ? $data['species_id'] : null;
                 $data['subtype_id'] = isset($data['subtype_id']) && $data['subtype_id'] ? $data['subtype_id'] : null;
+                $data['subtype_id_2'] = isset($data['subtype_id_2']) && $data['subtype_id_2'] ? $data['subtype_id_2'] : null;
                 $data['rarity_id'] = (isset($data['rarity_id']) && $data['rarity_id']) ? $data['rarity_id'] : null;
                 $data['transformation_id'] = isset($data['transformation_id']) && $data['transformation_id'] ? $data['transformation_id'] : null;
 
@@ -269,7 +278,7 @@ class CharacterManager extends Service
                 }
             }
             $imageData = Arr::only($data, [
-                'species_id', 'subtype_id', 'rarity_id', 'use_cropper',
+                'species_id', 'subtype_id', 'subtype_id_2', 'rarity_id', 'use_cropper',
                 'x0', 'x1', 'y0', 'y1', 'transformation_id',
             ]);
             $imageData['use_cropper'] = isset($data['use_cropper']) ;
@@ -648,6 +657,13 @@ class CharacterManager extends Service
                 if(!$subtype || $subtype->species_id != $data['species_id']) throw new \Exception('Selected subtype invalid or does not match species.');
             }
             else $data['subtype_id'] = null;
+            if(isset($data['subtype_id_2']) && $data['subtype_id_2'])
+            {
+                $subtypeTwo = Subtype::find($data['subtype_id_2']);
+                if(!(isset($data['species_id']) && $data['species_id'])) throw new \Exception('Species must be selected to select a subtype.');
+                if(!$subtypeTwo || $subtypeTwo->species_id != $data['species_id']) throw new \Exception('Selected secondary subtype invalid or does not match species.');
+            }
+            else $data['subtype_id_2'] = null;
 
             $data['is_visible'] = 1;
 
@@ -704,12 +720,19 @@ class CharacterManager extends Service
                 if(!(isset($data['species_id']) && $data['species_id'])) throw new \Exception('Species must be selected to select a subtype.');
                 if(!$subtype || $subtype->species_id != $data['species_id']) throw new \Exception('Selected subtype invalid or does not match species.');
             }
+            if(isset($data['subtype_id_2']) && $data['subtype_id_2'])
+            {
+                $subtypeTwo = Subtype::find($data['subtype_id_2']);
+                if(!(isset($data['species_id']) && $data['species_id'])) throw new \Exception('Species must be selected to select a subtype.');
+                if(!$subtypeTwo || $subtypeTwo->species_id != $data['species_id']) throw new \Exception('Selected secondary subtype invalid or does not match species.');
+            }
 
             // Log old features
             $old = [];
             $old['features'] = $this->generateFeatureList($image);
             $old['species'] = $image->species_id ? $image->species->displayName : null;
             $old['subtype'] = $image->subtype_id ? $image->subtype->displayName : null;
+            $old['subtype_2'] = $image->subtype_id_2 ? $image->subtypeTwo->displayName : null;
             $old['rarity'] = $image->rarity_id ? $image->rarity->displayName : null;
             $old['transformation'] = $image->transformation_id ? $image->transformation->displayName : null;
 
@@ -726,6 +749,7 @@ class CharacterManager extends Service
             // Update other stats
             $image->species_id = $data['species_id'];
             $image->subtype_id = $data['subtype_id'] ?: null;
+            $image->subtype_id_2 = $data['subtype_id_2'] ?: null;
             $image->rarity_id = $data['rarity_id'];
             $image->transformation_id = $data['transformation_id'] ?: null;
             $image->save();
@@ -734,6 +758,7 @@ class CharacterManager extends Service
             $new['features'] = $this->generateFeatureList($image);
             $new['species'] = $image->species_id ? $image->species->displayName : null;
             $new['subtype'] = $image->subtype_id ? $image->subtype->displayName : null;
+            $new['subtype_2'] = $image->subtype_id_2 ? $image->subtypeTwo->displayName : null;
             $new['rarity'] = $image->rarity_id ? $image->rarity->displayName : null;
             $new['transformation'] = $image->transformation_id ? $image->transformation->displayName : null;
 
@@ -1858,6 +1883,7 @@ is_object($sender) ? $sender->id : null,
                 'rarity_id' => $character->image->rarity_id,
                 'species_id' => $character->image->species_id,
                 'subtype_id' => $character->image->subtype_id,
+                'subtype_id_2' => $character->image->subtype_id_2,
                 'transformation_id' => $character->image->transformation_id
             ];
 
@@ -2123,12 +2149,16 @@ is_object($sender) ? $sender->id : null,
             if(isset($data['subtype_id']) && $data['subtype_id'])
                 $subtype = ($request->character->is_myo_slot && $request->character->image->subtype_id) ? $request->character->image->subtype : Subtype::find($data['subtype_id']);
             else $subtype = null;
+            if(isset($data['subtype_id_2']) && $data['subtype_id_2'])
+                $subtypeTwo = ($request->character->is_myo_slot && $request->character->image->subtype_id_2) ? $request->character->image->subtypeTwo : Subtype::find($data['subtype_id_2']);
+            else $subtypeTwo = null;
             if(isset($data['transformation_id']) && $data['transformation_id'])
                 $transformation = ($request->character->is_myo_slot && $request->character->image->transformation_id) ? $request->character->image->transformation : Transformation::find($data['transformation_id']);
             else $transformation = null;
             if(!$rarity) throw new \Exception("Invalid rarity selected.");
             if(!$species) throw new \Exception("Invalid ".__('lorekeeper.species')." selected.");
             if($subtype && $subtype->species_id != $species->id) throw new \Exception(ucfirst(__('lorekeeper.subtype'))." does not match the ".__('lorekeeper.species').".");
+            if($subtypeTwo && $subtypeTwo->species_id != $species->id) throw new \Exception("Secondary subtype does not match the species.");
 
             // Clear old features
             $request->features()->delete();
@@ -2155,6 +2185,7 @@ is_object($sender) ? $sender->id : null,
             $request->species_id = $species->id;
             $request->rarity_id = $rarity->id;
             $request->subtype_id = $subtype ? $subtype->id : null;
+            $request->subtype_id_2 = $subtypeTwo ? $subtypeTwo->id : null;
             $request->transformation_id = $transformation ? $transformation->id : null;
             $request->has_features = 1;
             $request->save();
@@ -2276,6 +2307,7 @@ is_object($sender) ? $sender->id : null,
                 'y1' => $request->y1,
                 'species_id' => $request->species_id,
                 'subtype_id' => ($request->character->is_myo_slot && isset($request->character->image->subtype_id)) ? $request->character->image->subtype_id : $request->subtype_id,
+                'subtype_id_2' => ($request->character->is_myo_slot && isset($request->character->image->subtype_id_2)) ? $request->character->image->subtype_id_2 : $request->subtype_id_2,
                 'transformation_id' => ($request->character->is_myo_slot && isset($request->character->image->transformation_id)) ? $request->character->image->transformation_id : $request->transformation_id,
                 'rarity_id' => $request->rarity_id,
                 'sort' => 0,
